move TestZipShouldRemove to provisioner_test.go 	put the test code together.
